ai_chatbot/requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.30,<1.0
pydantic>=2.6,<3.0
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0

